MOS-22842 : Updated deletion of pre-reg after packet creation
Former-commit-id: 23ec25e92204a07719b9aeeb26c2c08b5ca5fd66
Former-commit-id: 04134fc4a4837d8c2ae8f35a6a05f564d4644c0f

+++ registration/registration-client/src/main/java/io/mosip/registration/controller/reg/PacketHandlerController.java
@@ -650,14 +650,17 @@ public class PacketHandlerController extends BaseController implements Initializ
 				}
 				
 				//Deletes the pre registration Data after creation of registration Packet.
-				if(getRegistrationDTOFromSession().getPreRegistrationId()!=null) {
-					
+				if (getRegistrationDTOFromSession().getPreRegistrationId() != null
+						&& !getRegistrationDTOFromSession().getPreRegistrationId().isEmpty()) {
+
 					ResponseDTO responseDTO = new ResponseDTO();
 					List<PreRegistrationList> preRegistrationLists = new ArrayList<>();
-					PreRegistrationList preRegistrationList = preRegistrationDataSyncService.getPreRegistrationRecordForDeletion(getRegistrationDTOFromSession().getPreRegistrationId());
+					PreRegistrationList preRegistrationList = preRegistrationDataSyncService
+							.getPreRegistrationRecordForDeletion(
+									getRegistrationDTOFromSession().getPreRegistrationId());
 					preRegistrationLists.add(preRegistrationList);
 					preRegistrationDataSyncService.deletePreRegRecords(responseDTO, preRegistrationLists);
-					
+
 				}
 
 				// Generate the file path for storing the Encrypted Packet and Acknowledgement
